Made navbar responsive
fuck it, we're using bootstrap now.

look up bootstrap, it basically has a bunch of premade components that are all responsive that we can use.

The most important parts to know about it are related to its layouts which let us essentially copy and paste website layouts into our project and then style them.